+++ src/main/java/net/geforcemods/securitycraft/blocks/BlockMotionActivatedLight.java
@@ -1,5 +1,6 @@
 package net.geforcemods.securitycraft.blocks;
 
+import net.geforcemods.securitycraft.api.IModuleInventory;
 import net.geforcemods.securitycraft.tileentity.TileEntityMotionLight;
 import net.minecraft.block.Block;
 import net.minecraft.block.SoundType;
@@ -80,6 +81,16 @@ public class BlockMotionActivatedLight extends BlockOwnable {
 			dropBlockAsItem(world, pos, state, 0);
 			world.setBlockToAir(pos);
 		}
+	}
+
+	@Override
+	public void breakBlock(World world, BlockPos pos, IBlockState state) {
+		TileEntity te = world.getTileEntity(pos);
+
+		if (te instanceof IModuleInventory)
+			((IModuleInventory) te).dropAllModules();
+
+		super.breakBlock(world, pos, state);
 	}
 
 	@Override